{{cookiecutter.project_name}}/requirements.txt
boto3==1.26.51
django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
django-import-export==2.9.0
django-filter==22.1
django-model-utils==4.3.1
django-redis==5.2.0
django-simpleui==2022.12.18
django-storages==1.13.2
dj-database-url==1.2.0
drf-nested-routers==0.93.4
drf_yasg==1.21.4
psycopg2==2.9.5
redis==4.3.4
requests==2.27.1
sentry-sdk==1.13.0
uwsgi==2.0.20
